+++ src/main/java/blackjack/domain/Card.java
@@ -1,4 +1,4 @@
-package domain;
+package blackjack.domain;
 
 import java.util.Objects;
