+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,4 +1,4 @@
-package domain;
+package blackjack.domain;
 
 public class Dealer extends Gamer {
 	public boolean isHigher(Gamer gamer) {
