+++ src/main/java/blackjack/domain/Gamer.java
@@ -1,4 +1,4 @@
-package domain;
+package blackjack.domain;
 
 import java.util.ArrayList;
 import java.util.List;
