+++ src/main/java/blackjack/domain/Name.java
@@ -1,4 +1,4 @@
-package domain;
+package blackjack.domain;
 
 import java.util.regex.Pattern;
