+++ src/main/java/blackjack/domain/Player.java
@@ -1,4 +1,4 @@
-package domain;
+package blackjack.domain;
 
 import java.util.Objects;
