+++ src/main/java/blackjack/domain/Players.java
@@ -1,6 +1,7 @@
-package domain;
+package blackjack.domain;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 
 public class Players {
@@ -10,6 +11,7 @@ public class Players {
 		if (playerNames.isEmpty()) {
 			throw new IllegalArgumentException("플레이어를 한명 이상 입력해야 합니다.");
 		}
+		validateDuplicateName(playerNames);
 		players = new ArrayList<>();
 		for (String playerName : playerNames) {
 			players.add(new Player(playerName));
@@ -23,5 +25,11 @@ public class Players {
 	public void addCardToAllPlayers(Deck deck) {
 		players.stream()
 			.forEach(player -> player.addTwoCards(deck));
+	}
+
+	private void validateDuplicateName(List<String> playerNames) {
+		if (new HashSet<String>(playerNames).size() != playerNames.size()) {
+			throw new IllegalArgumentException("이름에 중복이 있으면 안됩니다.");
+		}
 	}
 }
