+++ src/main/java/blackjack/domain/Result.java
@@ -1,4 +1,4 @@
-package domain;
+package blackjack.domain;
 
 import java.util.HashMap;
 import java.util.List;
