+++ src/main/java/blackjack/domain/Type.java
@@ -1,4 +1,4 @@
-package domain;
+package blackjack.domain;
 
 public enum Type {
 	HEART("하트"),
