+++ src/main/java/blackjack/view/InputView.java
@@ -1,4 +1,4 @@
-package view;
+package blackjack.view;
 
 import java.util.Arrays;
 import java.util.List;
