+++ src/test/java/blackjack/domain/GamerTest.java
@@ -1,4 +1,4 @@
-package domain;
+package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.*;
