+++ src/test/java/blackjack/domain/ResultTypeTest.java
@@ -1,4 +1,4 @@
-package domain;
+package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.*;
