--- src/test/java/domain/PlayerTest.java
@@ -1,10 +1,0 @@
-package domain;
-
-import static org.assertj.core.api.Assertions.*;
-
-import org.junit.jupiter.api.Test;
-
-// public class PlayerTest {
-// 	@Test
-// 	void
-// }
